CALOPS-57: Menu rename — SITE / EVENTS / USER ACTIVITY structure
Restructure analytics nav per Toby-approved Read A (2026-05-12):
- SITE ACTIVITY  (top-level standalone link; visitor-heatmap)
- EVENTS ACTIVITY  (section): Creation, Audit Trail
- USER ACTIVITY  (section): Login, Search, Stream, Map & Center
- Errors  (kept as separate section, no longer collapse-nested)

Route move: /user-logins → /analytics/logins with server-side redirect
preserving old bookmarks.

Page headings updated to match menu hierarchy:
- "Visitor Heatmap" → "Site Activity"
- "Event Creation Analytics" → "Events Activity — Creation"
- "Event Activity Audit Trail" → "Events Activity — Audit Trail"
- "User Logins" → "User Activity — Login"
- "Unified User Activity" → "User Activity — Search"
- "Activity Tracking" → "User Activity — Stream"
- "Activity Map" → "User Activity — Map & Center"

AdminLayout.js: dropped Heatmaps/Tracking sections; replaced with new
4-section layout (Site Activity standalone + Events + User + Errors).
Removed unused errorsOpen state (no longer collapse).

FE-only, no BE changes.
