+++ src/components/AdminLayout.js
@@ -48,29 +48,31 @@ import StorageIcon from '@mui/icons-material/Storage';
 import MonetizationOnIcon from '@mui/icons-material/MonetizationOn';
 import MapIcon from '@mui/icons-material/Map';
 
-// Navigation items - Heatmaps section
-const heatmapNavItems = [
-  { name: 'Site Access', icon: <GridOnIcon />, href: '/dashboard/analytics/visitor-heatmap' },
-  { name: 'Event Creation', icon: <GridOnIcon />, href: '/dashboard/analytics/event-creation' },
+// Navigation — Site Activity (top-level standalone, no section)
+const siteActivityNav = {
+  name: 'Site Activity',
+  icon: <GridOnIcon />,
+  href: '/dashboard/analytics/visitor-heatmap',
+};
+
+// Navigation — Events Activity section
+const eventsActivityItems = [
+  { name: 'Creation',    icon: <GridOnIcon />, href: '/dashboard/analytics/event-creation' },
+  { name: 'Audit Trail', icon: <EventIcon />,  href: '/dashboard/analytics/event-activity' },
 ];
 
-// Navigation items - Tracking section (single item tracking)
-const trackingSingleItems = [
-  { name: 'Login Activity', icon: <AccessTimeIcon />, href: '/dashboard/user-logins' },
-  { name: 'Event Activity', icon: <EventIcon />, href: '/dashboard/analytics/event-activity' },
+// Navigation — User Activity section
+const userActivityItems = [
+  { name: 'Login',        icon: <AccessTimeIcon />,   href: '/dashboard/analytics/logins' },
+  { name: 'Search',       icon: <PersonSearchIcon />, href: '/dashboard/analytics/user-activity' },
+  { name: 'Stream',       icon: <AccessTimeIcon />,   href: '/dashboard/analytics/activity-tracking' },
+  { name: 'Map & Center', icon: <MapIcon />,          href: '/dashboard/analytics/activity-map' },
 ];
 
-// Navigation items - Tracking section (expanded search)
-const trackingExpandedItems = [
-  { name: 'User Activity', icon: <PersonSearchIcon />, href: '/dashboard/analytics/user-activity' },
-  { name: 'Activity Tracking', icon: <AccessTimeIcon />, href: '/dashboard/analytics/activity-tracking' },
-  { name: 'Activity Map', icon: <MapIcon />, href: '/dashboard/analytics/activity-map' },
-];
-
-// Navigation items - Tracking section (errors submenu)
-const trackingErrorItems = [
-  { name: 'API', icon: <BugReportIcon />, href: '/dashboard/errors/api' },
-  { name: 'Trends', icon: <TimelineIcon />, href: '/dashboard/errors/trends' },
+// Navigation — Errors section
+const errorsItems = [
+  { name: 'API',    icon: <BugReportIcon />, href: '/dashboard/errors/api' },
+  { name: 'Trends', icon: <TimelineIcon />,  href: '/dashboard/errors/trends' },
 ];
 
 // Navigation items - Data section
@@ -123,7 +125,6 @@ export default function AdminLayout({ children }) {
   const [anchorEl, setAnchorEl] = useState(null);
   const [appMenuAnchor, setAppMenuAnchor] = useState(null);
   const [organizerTypesOpen, setOrganizerTypesOpen] = useState(false);
-  const [errorsOpen, setErrorsOpen] = useState(false);
   const [gitOpen, setGitOpen] = useState(false);
 
   const handleDrawerToggle = () => {
@@ -191,12 +192,24 @@ export default function AdminLayout({ children }) {
 
       <Divider sx={{ my: 1 }} />
 
-      {/* Heatmaps Section */}
+      {/* Site Activity — standalone top-level link (no section header, only one page) */}
+      <List dense>
+        <ListItem disablePadding>
+          <ListItemButton component="a" href={siteActivityNav.href}>
+            <ListItemIcon>{siteActivityNav.icon}</ListItemIcon>
+            <ListItemText primary={siteActivityNav.name} />
+          </ListItemButton>
+        </ListItem>
+      </List>
+
+      <Divider sx={{ my: 1 }} />
+
+      {/* Events Activity */}
       <Typography variant="overline" sx={{ px: 2, pt: 1, color: 'text.secondary', fontWeight: 'bold' }}>
-        Heatmaps
+        Events Activity
       </Typography>
       <List dense>
-        {heatmapNavItems.map((item) => (
+        {eventsActivityItems.map((item) => (
           <ListItem key={item.name} disablePadding>
             <ListItemButton component="a" href={item.href}>
               <ListItemIcon>{item.icon}</ListItemIcon>
@@ -208,12 +221,12 @@ export default function AdminLayout({ children }) {
 
       <Divider sx={{ my: 1 }} />
 
-      {/* Tracking Section */}
+      {/* User Activity */}
       <Typography variant="overline" sx={{ px: 2, pt: 1, color: 'text.secondary', fontWeight: 'bold' }}>
-        Tracking
+        User Activity
       </Typography>
       <List dense>
-        {trackingSingleItems.map((item) => (
+        {userActivityItems.map((item) => (
           <ListItem key={item.name} disablePadding>
             <ListItemButton component="a" href={item.href}>
               <ListItemIcon>{item.icon}</ListItemIcon>
@@ -221,8 +234,16 @@ export default function AdminLayout({ children }) {
             </ListItemButton>
           </ListItem>
         ))}
-        <Divider sx={{ my: 0.5, mx: 2 }} />
-        {trackingExpandedItems.map((item) => (
+      </List>
+
+      <Divider sx={{ my: 1 }} />
+
+      {/* Errors */}
+      <Typography variant="overline" sx={{ px: 2, pt: 1, color: 'text.secondary', fontWeight: 'bold' }}>
+        Errors
+      </Typography>
+      <List dense>
+        {errorsItems.map((item) => (
           <ListItem key={item.name} disablePadding>
             <ListItemButton component="a" href={item.href}>
               <ListItemIcon>{item.icon}</ListItemIcon>
@@ -230,34 +251,6 @@ export default function AdminLayout({ children }) {
             </ListItemButton>
           </ListItem>
         ))}
-        <Divider sx={{ my: 0.5, mx: 2 }} />
-        {/* Errors submenu within Tracking */}
-        <ListItem disablePadding>
-          <ListItemButton
-            onClick={() => setErrorsOpen(!errorsOpen)}
-            onTouchEnd={(e) => {
-              e.preventDefault();
-              setErrorsOpen(!errorsOpen);
-            }}
-            sx={{ touchAction: 'manipulation', WebkitTapHighlightColor: 'transparent' }}
-          >
-            <ListItemIcon><BugReportIcon /></ListItemIcon>
-            <ListItemText primary="Errors" />
-            {errorsOpen ? <ExpandLess /> : <ExpandMore />}
-          </ListItemButton>
-        </ListItem>
-        <Collapse in={errorsOpen} timeout="auto" unmountOnExit>
-          <List component="div" disablePadding dense>
-            {trackingErrorItems.map((item) => (
-              <ListItem key={item.name} disablePadding>
-                <ListItemButton component="a" href={item.href} sx={{ pl: 4, touchAction: 'manipulation' }}>
-                  <ListItemIcon>{item.icon}</ListItemIcon>
-                  <ListItemText primary={item.name} />
-                </ListItemButton>
-              </ListItem>
-            ))}
-          </List>
-        </Collapse>
       </List>
 
       <Divider sx={{ my: 1 }} />
